# src/week2/instructor-code/src/week2/angular-demo/specifications/questions-openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Qa.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://qa_api.dev.localhost:9300/"
    }
  ],
  "paths": {
    "/questions/{id}": {
      "get": {
        "tags": ["Question"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/questions": {
      "get": {
        "tags": ["Question"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionListItem"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionListItem"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionListItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Question"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionSubmissionItem"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionSubmissionItem"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionSubmissionItem"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "QuestionListItem": {
        "required": ["id", "questionBody"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "maxLength": 100,
            "minLength": 5,
            "type": "string"
          },
          "questionBody": {
            "maxLength": 500,
            "minLength": 30,
            "type": "string"
          },
          "submittedAnswers": {
            "type": ["null", "array"],
            "items": {
              "$ref": "#/components/schemas/SubmittedAnswer"
            }
          }
        }
      },
      "QuestionSubmissionItem": {
        "required": ["title", "questionBody", "priority"],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 50,
            "minLength": 5,
            "type": "string"
          },
          "questionBody": {
            "maxLength": 500,
            "minLength": 30,
            "type": "string"
          },
          "priority": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          }
        }
      },
      "SubmittedAnswer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "questionBody": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Question"
    }
  ]
}